Extract scrolled navigation bar into widget type
REDMINE-19192
