+++ entry_types/scrolled/lib/pageflow_scrolled/plugin.rb
@@ -13,6 +13,10 @@ module PageflowScrolled
 
         c.revision_components.register(Storyline)
 
+        c.widget_types.register(ReactWidgetType.new(name: 'defaultNavigation',
+                                                    role: 'header'),
+                                default: true)
+
         c.features.register('datawrapper_chart_embed_opt_in')
         c.features.enable_by_default('datawrapper_chart_embed_opt_in')
       end
