+++ entry_types/scrolled/package/src/widgets/defaultNavigation/ChapterLink.module.css
