+++ entry_types/scrolled/package/src/widgets/defaultNavigation/DefaultNavigation.module.css
@@ -48,7 +48,7 @@
 }
 
 .contextIcon {
-  composes: unstyledButton from '../utils.module.css';
+  composes: unstyledButton from '../../frontend/utils.module.css';
   cursor: pointer;
   width: 40px;
   height: 50px;
