+++ entry_types/scrolled/package/src/widgets/defaultNavigation/HamburgerIcon.module.css
