+++ entry_types/scrolled/package/src/widgets/defaultNavigation/HamburgerIcons.module.css
