+++ entry_types/scrolled/package/src/widgets/defaultNavigation/LegalInfoLink.js
