+++ entry_types/scrolled/package/src/widgets/defaultNavigation/LegalInfoMenu.module.css
