+++ entry_types/scrolled/package/src/widgets/defaultNavigation/Logo.js
@@ -3,7 +3,7 @@ import classNames from 'classnames';
 
 import {useTheme} from 'pageflow-scrolled/frontend';
 
-import styles from './AppHeader.module.css';
+import styles from './DefaultNavigation.module.css';
 
 export function Logo() {
   const theme = useTheme();
