+++ entry_types/scrolled/package/src/widgets/defaultNavigation/Scroller.module.css
