+++ entry_types/scrolled/package/src/widgets/defaultNavigation/SharingMenu.module.css
