+++ entry_types/scrolled/package/src/widgets/defaultNavigation/SkipLinks.module.css
