+++ entry_types/scrolled/package/src/widgets/defaultNavigation/index.js
@@ -1,0 +1,6 @@
+import {frontend} from 'pageflow-scrolled/frontend';
+import {DefaultNavigation} from './DefaultNavigation';
+
+frontend.widgetTypes.register('defaultNavigation', {
+  component: DefaultNavigation
+});
